src/backend/services/api-service/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
httpx>=0.25.2
pytest>=7.4.0
email-validator>=2.0.0
python-jose>=3.3.0
motor>=3.3.2
pydantic
uvicorn
openai>=1.52.0
